Changed fragment classes' structure a little bit, utilized OOP for a hopefully more organized project structure.
